refactor(ui): break digital-twin-page.tsx into focused components #263
Split the 1954-line monolith into 15 focused files under
src/app/components/digital-twin/. Main page is now a 163-line
composition shell. No UI or behavioral changes.
