+++ src/app/components/digital-twin/health-score-gauge.tsx
@@ -1,0 +1,46 @@
+import { getHealthColor } from "./digital-twin-health-utils";
+
+// Story 15.1 — Circular Health Score Gauge
+export function HealthScoreGauge({ score, size = 80 }: { score: number; size?: number }) {
+  const strokeWidth = size > 60 ? 8 : 6;
+  const radius = (size - strokeWidth) / 2;
+  const circumference = 2 * Math.PI * radius;
+  const offset = circumference - (score / 100) * circumference;
+  const color = getHealthColor(score);
+
+  return (
+    <svg width={size} height={size} className="shrink-0">
+      <circle
+        cx={size / 2}
+        cy={size / 2}
+        r={radius}
+        fill="none"
+        stroke="#f1f3f5"
+        strokeWidth={strokeWidth}
+      />
+      <circle
+        cx={size / 2}
+        cy={size / 2}
+        r={radius}
+        fill="none"
+        stroke={color}
+        strokeWidth={strokeWidth}
+        strokeDasharray={circumference}
+        strokeDashoffset={offset}
+        strokeLinecap="round"
+        transform={`rotate(-90 ${size / 2} ${size / 2})`}
+        className="transition-all duration-500"
+      />
+      <text
+        x={size / 2}
+        y={size / 2}
+        textAnchor="middle"
+        dominantBaseline="central"
+        className="text-foreground font-bold"
+        style={{ fontSize: size > 60 ? 18 : 14 }}
+      >
+        {score}
+      </text>
+    </svg>
+  );
+}
